refactor: move selectors to separate file
