+++ src/redux/addresses/reducers.js
@@ -28,19 +28,4 @@ const addresses = createReducer(initialState, {
   },
 });
 
-export const getIsLoading = createSelector(
-  (state) => state.addresses.isLoading,
-  (isLoading) => isLoading,
-);
-
-export const getAddresses = createSelector(
-  (state) => state.addresses.list,
-  (list) => list,
-);
-
-export const getIsAddressesLoaded = createSelector(
-  (state) => state.addresses.list,
-  (list) => !!list.length,
-);
-
 export default addresses;
