+++ src/redux/route/selectors.js
@@ -1,0 +1,11 @@
+import { createSelector } from '@reduxjs/toolkit';
+
+export const getIsLoading = createSelector(
+  (state) => state.addresses.isLoading,
+  (isLoading) => isLoading,
+);
+
+export const getRoute = createSelector(
+  (state) => state.route.coords,
+  (coords) => coords,
+);
